2014/5 - 2020/21 Shot Data for the NBA's Denver Nuggets

A dataset containing seven seasons of Denver Nuggets shot data from NBA.com

nba_nuggets_shots

Format

A data frame with 1832 rows and 47 variables:

yearSeason

Season in which the game was played (YYYY)

slugSeason

Season in which the game was played (YYYY-YY)

idTeam

Team ID on nba.com

idPlayer

Player ID on nba.com

typeGrid

A string showing grid type

namePlayer

Player name

nameTeam

Full team name

typeEvent

Made or Missed Shot

typeAction

Type of shot, e.g. Jump Shot

typeShot

Two or three point shot

dateGame

Date in YYYY-MM-DD

slugTeamHome

Home Team abbreviation

slugTeamAway

Away Team abbreviation

idGame

Game ID on nba.com

idEvent

Event ID within the game on nba.com

numberPeriod

Period in game in which the shot was taken

minutesRemaining

Minutes remaining in the period in which the shot was taken

zoneBasic

Description of area on court where shot was taken

nameZone

Name of the zone

slugZone

Abbreviation of shot zone

zoneRange

Distance of shot in categories

locationX

X cordinate where shot was taken

locationY

Y cordinate where shot was taken

secondsRemaining

Seconds remainin when shot was taken

distanceShot

Distance of the shot in feet

isShotAttempted

Logical for shot taken

isShotMade

Logical for make/miss of shot

Source

http://asbcllc.com/nbastatR/reference/teams_shots.html